#55 - Design: Feedheader.jsx css 수정 및 header 파일 정리
